requirements.txt
google==3.0.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.29.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
huggingface-hub==0.34.3
numpy==2.3.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
qdrant-client==1.15.1
sentence-transformers==5.1.0
streamlit==1.48.0
tiktoken==0.11.0
tokenizers==0.21.4
transformers==4.55.0
newsapi-python
xmlrpc.client


